+++ src/util.rs
@@ -1,5 +1,5 @@
 use std::cmp::{min, max};
 
-pub fn clamp(value: usize, min_value: usize, max_value: usize) -> usize {
+pub fn clamp(value: i32, min_value: i32, max_value: i32) -> i32 {
     max(min_value, min(max_value, value))
 }
